Extracted seeding logic to common abstract test case
